cleanup from util functions
